RabiRamseySpectroscopy

Documentation for RabiRamseySpectroscopy.

RabiRamseySpectroscopy.FreeEvol

RabiRamseySpectroscopy.PerturbEvol

RabiRamseySpectroscopy.StateVector

RabiRamseySpectroscopy.probe

RabiRamseySpectroscopy.pump

RabiRamseySpectroscopy.pump!

RabiRamseySpectroscopy.wait

RabiRamseySpectroscopy.wait!

RabiRamseySpectroscopy.FreeEvol — Type

FreeEvol

Struct for a quantum physics operator that describes evolution of unperturbed quantum two-level system. This evolution can be described in the way of changing phase difference between |e> and |g> states.

Fields

operator: Represents operator itself in a form of a matrix.

Arguments

x : The phase difference between |e> and |g> states as a result of free evolution.

Returns

FreeEvol : Composite type instance.

See also

StateVector

References

Wikipedia: https://en.wikipedia.org/wiki/Two-statequantumsystem

RabiRamseySpectroscopy.PerturbEvol — Type

PerturbEvol

Struct for a quantum physics operator that describes evolution of quantum two-level system in a presence of some perturbation field.

Fields

operator: Represents operator itself in a form of a matrix.

Arguments

δ : The detuning, a measure of how far the perturbation field oscillation frequency is off-resonance relative to the transition (Hz).

τ : The duration of Rabi/Ramsey pulse (s).

Ω₀ : The Rabi frequency at which the probability amplitudes of two energy levels fluctuate in an oscillating perturbation field (Hz).

Returns

PerturbEvol : Composite type instance.

See also

StateVector

References

Wikipedia: https://en.wikipedia.org/wiki/Two-statequantumsystem

Wikipedia: https://en.wikipedia.org/wiki/Rabi_frequency

RabiRamseySpectroscopy.StateVector — Type

StateVector

Struct that describes a state of the quantum two-level system via ket vector in Dirac notation.

Fields

state : Vector that represents the state of the system in (|e>, |g>) basis.

Arguments

e : Numerical value that implies the probability of a quantum system to be in excited state or the amount of particles in the ground state. Value of |e> ket vector.

g : Numerical value that implies the probability of a quantum system to be in ground state or the amount of particles in the ground state. Value of |g> ket vector.

Returns

StateVector : Composite type instance.

References

Wikipedia: https://en.wikipedia.org/wiki/Two-statequantumsystem

Wikipedia: https://en.wikipedia.org/wiki/Bra%E2%80%93ket_notation

RabiRamseySpectroscopy.probe — Method

probe(state_vector::StateVector) -> Real

Method that computes the value of probability nₑ for a quantum system to be in excited state and can be interpret as population of excited state.

Arguments

self : The current state vector of a quantum system.

Returns

nₑ : Real number.

See also

RabiRamseySpectroscopy.StateVector

References

Wikipedia: https://en.wikipedia.org/wiki/Ramsey_interferometry

RabiRamseySpectroscopy.pump! — Method

pump!(self::StateVector, pulse::PerturbEvol) -> StateVector

Transform the state of a quantum system according to system's evolution in a presence of the laser pulse.

Arguments

self : The current state vector of a quantum system.

pulse : The laser pulse which is applied on the quantum system.

Returns

StateVector : Composite type instance with transformed state object.

See also

RabiRamseySpectroscopy.StateVector

RabiRamseySpectroscopy.PerturbEvol

References

Wikipedia: https://en.wikipedia.org/wiki/Ramsey_interferometry

RabiRamseySpectroscopy.pump — Method

pump(state_vector::StateVector, pulse::PerturbEvol) -> StateVector

Transform the state of a quantum system according to system's evolution in a presence of the laser pulse.

Arguments

state_vector : The current state vector of a quantum system.

pulse : The laser pulse which is applied on the quantum system.

Returns

StateVector : Composite type instance with transformed state object.

See also

RabiRamseySpectroscopy.StateVector

RabiRamseySpectroscopy.PerturbEvol

References

Wikipedia: https://en.wikipedia.org/wiki/Ramsey_interferometry

RabiRamseySpectroscopy.wait! — Method

wait!(self::StateVector, duration::FreeEvol) -> StateVector

Transform the state of a quantum system according to system's evolution in a presence of no perturbations.

Arguments

self : The current state vector of a quantum system.

duration : The time of free evolution of the quantum system.

Returns

StateVector : Composite type instance with transformed state object.

See also

RabiRamseySpectroscopy.StateVector

RabiRamseySpectroscopy.FreeEvol

References

Wikipedia: https://en.wikipedia.org/wiki/Ramsey_interferometry

RabiRamseySpectroscopy.wait — Method

wait(state_vector::StateVector, duration::FreeEvol) -> StateVector

Transform the state of a quantum system according to system's evolution in a presence of no perturbations.

Arguments

state_vector : The current state vector of a quantum system.

duration : The time of free evolution of the quantum system.

Returns

StateVector : Composite type instance with transformed state object.

See also

RabiRamseySpectroscopy.StateVector

RabiRamseySpectroscopy.FreeEvol

References

Wikipedia: https://en.wikipedia.org/wiki/Ramsey_interferometry